+++ src/enums/NewBlockMenuStyle.php
@@ -3,13 +3,13 @@
 namespace benf\neo\enums;
 
 /**
- * Pseudo-enum for representing new block button styles.
+ * Pseudo-enum for representing new block menu styles.
  *
  * @package benf\neo\enums
  * @author Spicy Web <plugins@spicyweb.com.au>
  * @since 3.6.0
  */
-abstract class NewBlockButtonStyle
+abstract class NewBlockMenuStyle
 {
     public const Classic = 'classic';
     public const Grid = 'grid';